requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
blinker==1.6.2
    # via flask
click==8.1.3
    # via flask
flask==2.3.2
    # via
    #   -r requirements.in
    #   flask-login
flask-login==0.6.2
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector-repackaged==0.3.1
    # via -r requirements.in
werkzeug==2.3.4
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
